[+] DTEngineArchitecture (cut from todos) [+] RenderingIssues with some thoughts about rendering architecture and solutions [+] Event class with delegates (no more usage of std::function a.k.a. Function) [*] Input reworked to match new Event system (instead of Functions) [*] DECLARE_SHARED_FROM_THIS macro which allows to avoid StaticPointerCast calls on shared_from_this by replacing it with SharedFromThis function [*] Changed std::enable_shared_from_this to EnableSharedFromThis and std::shared_ptr to SharedPtr in some places [*] ScreenToWorld and WorldToScreen points calculation improved [*] Changed index buffer type from uint64 to uint32 [*] Added some consts to local variables

+++ DTEngine/src/Rendering/MeshBase.cpp
@@ -15,7 +15,7 @@ MeshBase::~MeshBase()
 {
 }
 
-bool MeshBase::CreateBuffers(VertexType* vertices, uint64* indices)
+bool MeshBase::CreateBuffers(VertexType* vertices, uint32* indices)
 {
 	Graphics& graphics = GetGraphics();
 
@@ -40,7 +40,7 @@ bool MeshBase::CreateBuffers(VertexType* vertices, uint64* indices)
 	// Index buffer creation
 	// Filling desc
 	bufferDesc.Usage = D3D11_USAGE_DEFAULT;
-	bufferDesc.ByteWidth = sizeof(uint64) * _indicesCount;
+	bufferDesc.ByteWidth = sizeof(uint32) * _indicesCount;
 	bufferDesc.BindFlags = D3D11_BIND_INDEX_BUFFER;
 
 	// Filling data
